Implement proper SuperTask task-level splitting.
New SuperTask interface depends on a task-provided method
(makeTargetList) to split input data into individual data chunks to run
task on (with optional concurrent execution). Re-imlemented
ValidateDrpTask to use this new method, this means that execute()
method (and run()) will be called now for single-filter data. The logic
to split data into per-filter sets was moved from run() method to
makeTargetList().

Note that this will going to change in the future again once we determine
what is the best design for SuperTask is.